+++ src/shared/components/Drawer/components/Aside/components/DialogNewBoard/DialogNewBoardView.tsx
@@ -15,7 +15,7 @@ import { IDialogNewBoardView } from "shared/components/Drawer/components/Aside/c
 import { BOARD_TYPE_ID } from "shared/constants";
 import { FiX } from "react-icons/fi";
 import { ICON_SIZE } from "shared/constants";
-import { fieldErrors } from "shared/utils/fieldErrors";
+import { fieldErrors } from "shared/utils/field-errors";
 
 const DialogNewBoardView: React.FC<IDialogNewBoardView> = props => {
 	const {
