+++ src/shared/components/TextFieldPassword/TextFieldPassword.tsx
@@ -4,7 +4,7 @@ import { FiEye, FiEyeOff } from "react-icons/fi";
 import { ICON_SIZE } from "shared/constants";
 import { IconButton, TextFieldPasswordContainer } from "./style";
 import { ITextFieldPassword } from "./types";
-import { fieldErrors } from "shared/utils/fieldErrors";
+import { fieldErrors } from "shared/utils/field-errors";
 
 const TextFieldPassword: React.FC<ITextFieldPassword> = forwardRef(
 	({ register, errors }, ref) => {
