--- src/shared/utils/fieldErrors/index.ts
@@ -1,1 +1,0 @@
-export * from "./fieldErrors";
